simplified mcaanalysiswindow progressBar
cleaned up imports in pptaskmanagers

removed call to gc.collect on pptaskmanager timeout. We dont want the additional
overhead unless we can prove it is absolutely necessary, in which case we can
add a call during report

set the scanData queue in pptaskmanager.setData. This is a temporary cleanup,
the next step is to improve the iteration by moving all the logic into the
scanData, maybe using generators and avoiding the queues altogether.