implemented pptaskmanager based on data iterators. McaSpectraIterator skips invalid points, uses exceptions to indicate when we are waiting for spec to report subsequent points and when we have reached the end of the data.
parallelpython reports that each job takes about 0.8-0.9 seconds to complete. 0.010 seconds
is used to call estimate(), about 0.2 seconds on average for fit(), and .005 seconds for
concentrations. So it seems 0.6-0.7 seconds is consumed in overhead, probably for the 
required imports, pickle and unpickle steps. It should not be too difficult to implement
an alternative taskmanager based ipython taskfarming, which shouldnt require as much
overhead.