Large scale changes in handling diffraction detector specs, new background algorithm, new data scaling and dezing operations.
Due to new GE "area_detector" a lot of flexibility with respect to detector parameters has been added:

Detectors are still stored as square arrays. Circle and rectangles must be "grown" to the smallest square that inscribes them.

There are dataset-specific settings for "psize" and "bin". "psize is the detector mixel pitch in mm. Concerning bin: a lot of visualization and some math is performed on binned versions of the XRD data. Binned data is an average over bin x bin pixels, where the detector width must be divisible by bin.

Integer math was previoiusly performed with uint16 but now the appropriate datatype for the diffraction images is acertained and used.

For the first time, there are routines that MODIFY RAW DATA. Previously, any data in the '/<scan>/measurement' group of an h5 file was not touched, but now there is one GUI that can call 4 math functions that modify '/<scan>/measurement/<detector>/counts' and give it attributes that describe the modification:
. dezing: This is to get rid of zingers in CCD images. This happens either by looking for saturated pixels or by looking for pixels with values much bigger than all 4 of the neighbors
. normalization: Any array in '/<scan>/measurement/scalar_data' can be used to normalize the diffraction data
. scaling: If normalizing with a quantity far from unity, the result will not be ammenable to saving with the same datatype (e.g. if using ion counter normalization) so there is also a factor you can specify and 1/scalar_data array will be mutliplied by this factor before normalizing the counts array


There is a new 2-d background subtraction capability call "lin" in which the user specifies 2 diffraction images and for each diffraction image in the dataset, the a background images is calculated as a linear combination of the 2 images. The "weights" of the two images are determined by maximizing the total umber of counts in the background image under the contraint that the background exceeds the data at only a small fraction of pixels. The 2 images specified for the linear background can come from within the present dataset or from an inventory stored in the attrs_maps.h5 file.